Rename core submodule files to core.py
