+++ pythainlp/corpus/core.py
