+++ pythainlp/spell/__init__.py
@@ -14,4 +14,4 @@ from pythainlp.spell.pn import NorvigSpellChecker
 
 DEFAULT_SPELL_CHECKER = NorvigSpellChecker()
 
-from pythainlp.spell.spell import correct, spell
+from pythainlp.spell.core import correct, spell
