+++ pythainlp/spell/core.py
