+++ pythainlp/tokenize/core.py
